+++ src/components/cards/CardFull/index.tsx
