+++ src/components/cards/CardFull/styles.module.scss
